backend/requirements.txt
flask
flask-cors
tensorflow
keras
numpy
pandas
scikit-learn
python-dotenv
pymongo
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai
PyJWT
